调整游戏数据架构
