+++ Source/StoneDefence/Core/GameCore/TowersDefencePlayerState.h
@@ -4,7 +4,13 @@
 
 #include "CoreMinimal.h"
 #include "../RuleOfThePlayerState.h"
+#include "../../Data/Save/PlayerSaveData.h"
+#include "../../Data/PlayerData.h"
+#include "../../Data/BuildingTowerData.h"
 #include "TowersDefencePlayerState.generated.h"
+
+
+
 
 /**
  * 
@@ -14,8 +20,21 @@ class STONEDEFENCE_API ATowersDefencePlayerState : public ARuleOfThePlayerState
 {
 	GENERATED_BODY()
 	
+public:
+	ATowersDefencePlayerState();
 
+	FPlayerData &GetPlayerData();
+	FBuildingTower &GetBuildingTower(const FGuid &ID);
+	const TArray<const FGuid*> GetBuildingTowersID();
+	const FBuildingTower &AddBuildingTower(const FGuid &ID, const FBuildingTower &Data);
+	void RequestInventorySlotSwap(const FGuid &A, const FGuid &B);
+	FBuildingTower &GetBuildingDataNULL();
+	UPlayerSaveData *GetSaveData();
 protected:
 	UPROPERTY()
 		UPlayerSaveData *SaveData;
+private:
+
+	UPROPERTY()
+		FBuildingTower BuildingTowerNULL;
 };
