+++ Source/StoneDefence/Data/Save/PlayerSaveData.h
@@ -3,7 +3,12 @@
 #pragma once
 
 #include "CoreMinimal.h"
+#include "../Core/SaveGameCore.h"
+#include "../PlayerData.h"
+#include "GameSaveData.h"
 #include "PlayerSaveData.generated.h"
+
+
 
 
 
@@ -16,53 +21,19 @@ class STONEDEFENCE_API UPlayerSaveData : public USaveGameCore
 	GENERATED_BODY()
 
 public:
-	UPlayerSaveData();
-
-	void Init();
-
-	bool IsValid();
-
-	//玩家ID
+	//玩家数据
 	UPROPERTY(SaveGame)
-		int32 PlayID;
+	FPlayerData PlayerData;
 
-	//玩家名字
+	//背包数据
 	UPROPERTY(SaveGame)
-		FName PlayName;
+	TMap<FGuid, FBuildingTower> BuildingTowers;
 
-	//队伍
-	UPROPERTY(SaveGame)
-		bool bTeam;
+	////玩家游戏中数据的状态
+	//UPROPERTY(SaveGame)
+	//TMap<FGuid, FPlayerSkillData> PlayerSkillDatas;
 
-	//玩家账户
-	UPROPERTY(SaveGame)
-		FString Account;
+	//virtual void InitSaveGame(UWorld *InWorld);
 
-	//游戏金币
-	UPROPERTY(SaveGame)
-		int32 GameGold;
-
-	//钻石 需要购买
-	UPROPERTY(SaveGame)
-		int32 Diamonds;
-
-	//铜币 
-	UPROPERTY(SaveGame)
-		int32 Copper;
-
-	UPROPERTY(SaveGame)
-		float GameGoldTime;
-
-	UPROPERTY(SaveGame)
-		float MaxGameGoldTime;
-
-	//玩家拥有的技能
-	UPROPERTY(SaveGame)
-		TArray<int32> SkillIDs;
-
-	//玩家游戏中数据状态
-	UPROPERTY(SaveGame)
-		TArray<FPlayerSkillData> SkillData;
-
-	bool IsAllowIncrease();
+	//void AddPlayerSkill(UWorld *InWorld, const FGuid *Guid, int32 SkillID);
 };
